Adding in the beginnings of an idea I'm toying with to pass in the type of data some block of data actually is via a table in an external file.
Currently, the function and enum don't really do anything but they may in the future.

The idea is to have a 2 column table with the left side consisting of an identifier and the right side consisting of the type of data to assume for the left side and then set a type variable for the blocks that correspond to the identifiers with the proper data type.

The idea for the end result is that it'd be easy to convert to a specific type of variable for output or something just by checking the type variable for some data block.

It's still being toyed with, though, and I haven't really decided on how I want to handle the table and stuff yet. /shrug