Fleshing out the reading of the group stuff.
Compression handling is still left to be done for the record stuff.

Also adding a note to the TODO list about refactoring and adding in a dependency on parser.cpp for fileFormat.cpp, which may be temporary (haven't decided yet), to reuse the GRUP checking routine that was already written.

The logic still needs to be checked out for all of this, however.