Breaking out the reading of the file header into its own function.
This way, we do not need to read in the entire file each time we want just the header. This adds more flexibility as there are cases when just a list of masters is desired, so one could just call the header reading function.

It also helps make the code cleaner.