Refactor Record class
Make members private where appropriate, don't read the record type when
it's not used, and don't store the total fields size as a member
variable when it's only used during parsing.

Also split off field parsing into a separate function, and don't use a
counter to track bytes read after each and every read, as that's more
likely to go wrong than just reading the stream position when needed.

For field types, don't use a char array as that's trickier to work with:
read directly into a string instead.