Fix reading non-ASCII plugin paths
This needed three distinct fixes:
1. To stop MSVC mangling UTF-8 string literals, the UNICODE and _UNICODE
preprocessor definitions need to be set.
2. To get Boost Filesystem path objects to convert from UTF-8 to the
correct system encoding (UTF-16), the objects need to be imbued with a
locale that has a UTF-8/16 converter, otherwise the non-ASCII characters
are converted as ASCII gibberish, and the wrong file gets passed to the
system functions internally that do the actual creating (see the file
copy added).
2. At least on Windows, std::fstream can't open UTF-8 file paths. I'm
not sure exactly where it's going wrong, but if I use Boost Filesystem's
overloaded file streams and pass file objects directly, the problem goes
away.