Fix incorrect group size used for Oblivion plugins
This causes a parsing error for the Knights.esp DLC plugin among others,
but unfortunately I can't figure how to replicate the issue in a testing
plugin.

The size was 4 bytes too small, so would cause the group parser to exit
4 bytes too early, except that the parser reads what's actually in the
group, so would appear to go "past the end" of the group instead. I
can't think of the circumstances under which this would be a problem,
and Knights.esp is too big to trace execution through to see where
exactly it goes wrong.